requirements.txt
pandas
langchain-ollama
langchain-community
python-dotenv
streamlit
fastapi 
uvicorn 
python-jose[cryptography] 
passlib[bcrypt]
python-multipart
